Initial Commit
First commit of my first go package, Gonf! In the never ending theme of
YACL (yet another config loader) I have created this package that will
load a text file from the given path and attempt to parse it for key
value pairs and return a map[string]string of the results.

Comments are included using '#' at the start of lines. Inline commenting
is not yet supported but it will be! As will the possible functionality
of returning map with the values parsed into their respective types.
...Maybe.